backend/pyproject.toml
[project]
name = "ca-ai-backend"
version = "0.1.0"
description = "CA AI MVP Backend - Local Processing Engine"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "protobuf>=4.0.0",
    # OCR and Image Processing
    "paddleocr>=2.7.0",
    "paddlepaddle>=3.0.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "pdf2image>=1.16.0",
    # Document Parsing
    "pdfplumber>=0.10.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    # Embeddings and Vector Search
    "sentence-transformers>=2.7.0",
    "numpy>=1.24.0",
    # LLM Providers
    "anthropic>=0.75.0",
    "google-genai>=1.0.0",
    "groq>=0.37.0",
    "openai>=1.0.0",
    # Environment Variables
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["core", "services", "rules", "api"]

